+++ tests/testthat/manual_tests/coverage.R
@@ -1,5 +1,0 @@
-x <- testthis::get_test_coverage()
-
-x
-
-as.data.frame(x)